[QCF-24]  TitleWithInput 리펙토링
